+++ CHANGES.rst
@@ -5,6 +5,8 @@
 - Remove ``with_units`` argument from ``__call__`` and ``invert``. Use the Shared API
   functions instead, ``pixel_to_world`` and `world_to_pixel``. [#635]
 
+- Adjust and fix tests. Replace ``logging`` with `warnings``. [#659]
+
 0.26.1 (2025-11-19)
 -------------------
 - Fix an indexing bug in ``spectroscopy.SellmeierZemax`` where the output ``n`` for array-type wavelength